Extract functions canAccept, step from type class
The functions were exported from the module;
being part of the type class means that the consumer of the module
would be able to override their implementation.

It could be argued that function 'nextState' should also be extracted,
but since it's not exported from the module,
the consumer cannot override it.